Harden native allocation cleanup

+++ src/mat_vec.c
@@ -56,12 +56,12 @@ double **alloc_matd(int nrows, int ncols)
   
   if(ncols != 0) {
 
-    if((m=(double**)malloc(sizeof(double*)*ncols))==NULL) {
+    if((m=(double**)malloc(sizeof(double*) * (size_t)ncols))==NULL) {
       error("\nFATAL ERROR: Memory allocation failure (type DBL_MATRIX). Program terminated.\n");
     }
 
     for(i=0;i<ncols;i++) {
-      if((m[i]=(double*)malloc(sizeof(double)*nrows))==NULL) {
+      if((m[i]=(double*)malloc(sizeof(double) * (size_t)nrows))==NULL) {
         free_mat(m, i);
         error("\nFATAL ERROR: Memory allocation failure (type DBL_MATRIX). Program terminated.\n");
       }
